requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters

langchain-huggingface
langchain_google_genai
chromadb

sentence-transformers

python-dotenv